mechanical-nlp/requirements.txt
transformers
peft
torch
sentence-transformers
faiss-cpu
scikit-learn
pandas
numpy
tqdm
evaluate
datasets 